fixed overlays
