--- client/src/components/map/partials/Overlay.tsx
@@ -1,11 +1,0 @@
-import { FC } from "react";
-
-const Overlay: FC = ({ children }) => {
-  return (
-    <div className={"flex justify-between m-2 h-4/5 max-h-4/5"}>
-      {children}
-    </div>
-  );
-};
-
-export default Overlay;
